modify the config to load from environment vars
This encourages people to keep their sensitive config information out of
source control. By default, it will fall back to the 2nd parameter values
in the `env` method (available out of the box in Laravel 5+
via [phpdotenv](https://github.com/vlucas/phpdotenv)) which allows it to
function as basic documentation if a user decides not to put this info in
their `.env` file.